+++ src/matrices/mod.rs
@@ -1508,9 +1508,9 @@ where
  * Third party fixed precision and infinite precision decimal types should
  * be able to implement all of the methods for [Real] and then utilise these functions.
  */
-impl<T: Numeric + Real> Matrix<T>
+impl<T: Real> Matrix<T>
 where
-    for<'a> &'a T: NumericRef<T> + RealRef<T>,
+    for<'a> &'a T: RealRef<T>,
 {
     /**
      * Computes the [L2 norm](https://en.wikipedia.org/wiki/Euclidean_vector#Length)